Adds stackable support for Jump actions
Changes a bit the stackable support for Move actions:
 - Instead of using previousTick to compute the new position (which is
   fast, but can lead to precision issues),
 it uses a diff between the current position and the previous position.

 This new way of doing it is slower, but it is precise.

+++ cocos2d/ccTypes.h
@@ -381,7 +381,7 @@ typedef enum {
 
 //! delta time type
 //! if you want more resolution redefine it as a double
-typedef float ccTime;
+typedef CGFloat ccTime;
 //typedef double ccTime;
 
 typedef float ccMat4[16];
